Move Template Extraction to Engines
In order to better allow our pattern_assembler to be engine agnostic. It makes
sense to split this functionality out. When supporting different engines, there
are possibilities that the engine may not need a raw output of the file, but
the module exports instead (e.g. React). Thanks to the tests, this was able to
be refactored quite simply.