Add Initial React Engine Work
Currently, this will render static markup composed from React components. The single
test right now is for an individual component.

Nested components are going to need to be thought out. Adhering to the patlab spec
and assuring that we can provide a way for users to know what a pattern is composed
of will make implementation a little different than for previous engines.

**Necessary Deps** - *Developers will need these to run React engine*

`Babel` - currently the best and most reliable way to require in JSX, which we should
be flexible enough to account for.

`React` - the react library is needed to make components

`ReactDOM` - we are currently using it to render the static markup.

** Temp Deps** - *Developers will need these for now to run React engine*

`Lodash` - This can be removed, just for ease of use at the moment, I noticed the utilities
method that did something similar, but no tests for it.

**Note** - The addition, and placement of loadCompiledTemplates is by no means
permanent. The thought process behind this is that for React components, you are
not going to need the raw output of the file, you're going to need the module exports
from the file itself. This was the intention of externalizing the template call, that
way you could provide both, or one or the other.

If names and locations are not up to code, I'll be more than happy to adjust.