Clarify Tests, Add Nested Component Hydration Test
The tests now read a little cleaner, and are using default props. This will allow
us to test the `engine_react` portion a bit better due to the way the object assign
is working. Since we want to be able to override JSON data with props passed directly
to the Component, much like in the mustache/handlebars engines, we have the JSON props
object as the inner most assign.This allows default/JSON data to be overwritten by
this.props.